checkpointing for a move from cafe back to home
this commit fixes the bug with downloading the stupid install scripts,
in which the scripts call 'sudo' but cloud-init won't allow that to be
used since cloud-init doesn't technically assign a terminal to the
configuration process. Piping the scripts to 'sed' to remove the 'sudo'
lines seems to have worked, since the scripts are run by root in
single-user mode anyway.