checkpoint - new download URL for rexray, cleanups
the download endpoint for rexray has undergone a bunch of changes over
the past day or so, and has finally settled into a stable state.

the template is now valid enough that it can be used with the
'update-stack' command to update a running stack without errors... not
that we'd ever actually want to do that, but it's a good metric of
validity. the IAM creds are creating and populating the outputs, so next
step is to populate the rexray configs and test.